requirements.txt
pre-commit
pytest
pytest-cov
grpcio-tools
jsonschema
pyyaml
hvac
watchdog
pytest-asyncio
